+++ test_package_a3.py
@@ -1,0 +1,9 @@
+from ogs.core.logger import configure_logger
+
+configure_logger()
+
+from ogs.core.startup import StartupManager
+
+startup = StartupManager()
+
+startup.start()